Set campaign query params via JS from current URL
This works with caching and doesn't override campaign tag when coming from outside
